ninth commit

+++ database/migrations/2025_05_19_173343_create_mpesa_transactions_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('checkoutRequestId')->nullable();
             $table->string('mpesaReceiptNumber')->nullable();
             $table->string('phoneNumber')->nullable();
-            $table->decimal('amount', 10, 2)->nullable();
+            $table->integer('amount')->nullable();
             $table->dateTime('transactionDate');
             $table->timestamps();
